Change main plugin class name
